refactor: make util directory
